requirements.txt
# Core LangChain packages
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-huggingface==0.0.3

# Vector store
chromadb==0.5.5

# PDF parsing
pypdf==4.3.1

# Embeddings (free fallback)
sentence-transformers==3.0.1

# Token counting (needed by LangChain OpenAI)
tiktoken==0.7.0

# Web UI
streamlit==1.37.1

# Environment variable management
python-dotenv==1.0.1
